+++ src/main/java/dev/ftb/mods/ftbic/util/MachineConfiguration.java
@@ -7,14 +7,18 @@ import net.neoforged.neoforge.fluids.SimpleFluidContent;
 
 import java.util.List;
 
-public record MachineConfiguration(Identifier machine, SideConfiguration sides, List<GhostItem> inputLocks, List<GhostItem> batch, SimpleFluidContent batchFluid) {
+public record MachineConfiguration(Identifier machine, SideConfiguration sides, List<GhostItem> inputLocks, List<GhostItem> batch, SimpleFluidContent batchFluid, QuarryFilter quarryFilter) {
 	public MachineConfiguration {
 		inputLocks = List.copyOf(inputLocks);
 		batch = List.copyOf(batch);
 	}
 
 	public MachineConfiguration(Identifier machine, SideConfiguration sides) {
-		this(machine, sides, List.of(), List.of(), SimpleFluidContent.EMPTY);
+		this(machine, sides, List.of(), List.of(), SimpleFluidContent.EMPTY, QuarryFilter.DEFAULT);
+	}
+
+	public MachineConfiguration(Identifier machine, SideConfiguration sides, List<GhostItem> inputLocks, List<GhostItem> batch, SimpleFluidContent batchFluid) {
+		this(machine, sides, inputLocks, batch, batchFluid, QuarryFilter.DEFAULT);
 	}
 
 	public static final Codec<MachineConfiguration> CODEC = RecordCodecBuilder.create(i -> i.group(
@@ -22,6 +26,7 @@ public record MachineConfiguration(Identifier machine, SideConfiguration sides, 
 			SideConfiguration.CODEC.fieldOf("sides").forGetter(MachineConfiguration::sides),
 			GhostItem.LIST_CODEC.optionalFieldOf("input_locks", List.of()).forGetter(MachineConfiguration::inputLocks),
 			GhostItem.LIST_CODEC.optionalFieldOf("batch", List.of()).forGetter(MachineConfiguration::batch),
-			SimpleFluidContent.CODEC.optionalFieldOf("batch_fluid", SimpleFluidContent.EMPTY).forGetter(MachineConfiguration::batchFluid)
+			SimpleFluidContent.CODEC.optionalFieldOf("batch_fluid", SimpleFluidContent.EMPTY).forGetter(MachineConfiguration::batchFluid),
+			QuarryFilter.CODEC.optionalFieldOf("quarry_filter", QuarryFilter.DEFAULT).forGetter(MachineConfiguration::quarryFilter)
 	).apply(i, MachineConfiguration::new));
 }